\name{quartCluster}
\alias{quartCluster}
\title{
Quarternary Protein Structure Clustering Analysis
}
\description{
Perform the clustering analysis over the protein quarternary structure. One can select to use the \pkg{iPAC}, \pkg{GraphPAC} or \pkg{SpacePAC} methods. The output will show the relevant information from each algorithm that was selected.

}
\usage{
quartCluster(mutation_data, alignment, perform.ipac = "Y", perform.graphpac = "Y",
			 perform.spacepac = "Y", insertion.type = "cheapest_insertion", 
			 MultComp = "Bonferroni", alpha = 0.05, show.low.level.messages = "N", 
			 ipac.method = "MDS", spacepac.method = "SimMax", create.map = "Y",
			 Show.Graph = "Y", Graph.Output.Path = NULL, Graph.File.Name = "Map.pdf",
			 Graph.Title = "Mapping", fix.start.pos = "Y", numsims = 1000, 
			 simMaxSpheres = 3, radii.vector = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10), 
			 OriginX = "", OriginY = "", OriginZ = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mutation_data}{
The mutation data in the format outputted by \code{getMutations}. 
}
  \item{alignment}{
The assembly structural information outputted by \code{makeAlignedSuperStructure}.
}
  \item{perform.ipac}{
Whether or not to perform the \pkg{iPAC} algorithm. Either a "Y" or a "N".
}
  \item{perform.graphpac}{
Whether or not to perform the \pkg{GraphPAC} algorithm. Either a "Y" or a "N".
}
  \item{perform.spacepac}{
Whether or not to perform the \pkg{SpacePAC} algorithm. Either a "Y" or a "N".
}
  \item{insertion.type}{
Specifies the type of insertion method used in the \pkg{GraphpAC} package. Please see the \pkg{GraphPAC} for more details.}

  \item{MultComp}{
Specifies the multiple comparison adjustment required by the \pkg{iPAC} and \pkg{GraphPAC} packages. Options are: "Bonferroni", "BH", or "None". Please see the \pkg{iPAC} and \pkg{GraphPAC} packages for details.

}
  \item{alpha}{
The significance level required in order to find a mutational cluster significant using the \pkg{iPAC} and \pkg{GraphPAC} algorithms.
}
  \item{show.low.level.messages}{
Whether to display the output messages generated by the \pkg{iPAC}, \pkg{GraphPAC} and \pkg{SpacePAC} algorithms. Either a "Y" or a "N". Commonly used for debugging.
}
  \item{ipac.method}{
The type of approach used by \pkg{iPAC} to map the protein to 1D space. This parameter usually set to "MDS", but can be set to "linear" as well. See the \pkg{iPAC} package for more details.
}
  \item{spacepac.method}{
The type of approach used by \pkg{SpacePAC} to identify clustering. The options are either "SimMax" or "Poisson. This parameter usually set to "SimMax". See the \pkg{SpacePAC} package for more details.
}
  \item{create.map}{
Whether a graphical representation of the \pkg{iPAC} algorithm's dimension reduction from 3D to 1D space should be diplsayed. Either "Y" or "N".
}
  \item{Show.Graph}{
Whether to show the \pkg{iPAC} package dimension reduction chart on the screen. Warning: You must be running R in a GUI environment, otherwise, an error will occur.
}
  \item{Graph.Output.Path}{
Where to save the dimension reduction chart. This is useful if you want to save the chart automatically or can't display it on the screen (for instance, you are running R in a terminal window).The Graph.File.Name variable must be set as well.
}
  \item{Graph.File.Name}{
If you would like the chart saved automatically to the disk, specify the output file name. The Graph.Output.Path variable must be set as well.
}
  \item{Graph.Title}{
The title of the graph to be created.}
  \item{fix.start.pos}{
For the \pkg{GraphPAC} package, the heuristic solver for the traveling salesman problem starts the path at a random amino acid. In order to make the results easily reproducible, the default starts the path on the first amino acid in the protein. Please see the \pkg{GraphPAC} package for more details.
}
  \item{numsims}{
The number of times to simulate the distribution of mutations over the protein quarternary structure for the \pkg{SpacePAC} algorithm. For each simulation, given m total mutations and n total amino acids, each amino acid has a m/n probability of mutation.
}
  \item{simMaxSpheres}{
For the \pkg{SpacePAC} algorithm, the maximum number of spheres to consider. Currently, the implementation allows for simMaxspheres to be either 1, 2 or 3.
}
  \item{radii.vector}{
This applies to the \pkg{SpacePAC} algorithm and denotes the vector of radii that will be considered. At each sphere radius, the best sphere combination is found. See the \emph{SpaceClust} method in the \pkg{SpacePAC} package for further details}
  \item{OriginX}{
If the "Linear" method is chosen for the \pkg{iPAC} algorithm, this specifies the x-coordinate part of the fixed point. See the vignette in the \pkg{iPAC} package for more details.}
  \item{OriginY}{
If the "Linear" method is chosen for the \pkg{iPAC} algorithm, this specifies the y-coordinate part of the fixed point. See the vignette in the \pkg{iPAC} package for more details.
}
  \item{OriginZ}{
If the "Linear" method is chosen for the \pkg{iPAC} algorithm, this specifies the z-coordinate part of the fixed point. See the vignette in the \pkg{iPAC} package for more details.
}
}

\value{
\item{ipac}{The clustering results using the \pkg{iPAC} algorithm. See the \pkg{iPAC} packagee for more details of each sub item.}
\item{graphpac}{The clustering results using the \pkg{GraphPAC} algorithm. See the \pkg{GraphPAC} package for more details of each sub item.}
\item{spacepac}{The clustering results using the \pkg{SpacePAC} algorithm. See the \pkg{SpacePAC} package for more details of each sub item.}
\item{ipac_messages}{Any messages that might of been reported by the \pkg{iPAC} algorithm. Typically, warning or error messages are displayed here.}
\item{graphpac_messages}{Any messages that might of been reported by the \pkg{GraphPAC} algorithm. Typically, warning or error messages are displayed here.}
\item{spacepac_messages}{Any messages that might of been reported by the \pkg{SpacePAC} algorithm. Typically, warning or error messages are displayed here.}
}
\references{
Gregory Ryslik and Hongyu Zhao (2012). iPAC: Identification of Protein Amino acid Clustering. R package version 1.8.0.
Gregory Ryslik and Hongyu Zhao (2012). GraphPAC: Identification of Mutational Clusters in Proteins via a Graph Theoretical Approach.. R package version 1.6.0.
Gregory Ryslik and Hongyu Zhao (2013). SpacePAC: Identification of Mutational Clusters in 3D Protein Space via Simulation.. R package version 1.2.0.
Michael Hahsler and Kurt Hornik (2014). TSP: Traveling Salesperson Problem (TSP). R package version 1.0-9. http://CRAN.R-project.org/package=TSP
}
\note{
The clustering results give the \emph{serial number} values from the *.pdb1 file. 

Most of the parameters simply pass the requisite values to the underlying \pkg{iPAC}, \pkg{GraphPAC} and \pkg{SpacePAC} algorithms. The user should be aware of the parameters for these algorithms as this package is designed to extend them to quarternary  structures.
}
\examples{
\dontrun{
#read the mutational data
mutation_files <- list(
        system.file("extdata","HFE_Q30201_MutationOutput.txt", package = "QuartPAC"),
        system.file("extdata","B2M_P61769_MutationOutput.txt", package = "QuartPAC")
					)
uniprots <- list("Q30201","P61769")
mutation.data <- getMutations(mutation_files = mutation_files, uniprots = uniprots)

#read the pdb file
pdb.location <- "http://www.rcsb.org/pdb/files/1A6Z.pdb"
assembly.location <- "http://www.rcsb.org/pdb/files/1A6Z.pdb1"
structural.data <- makeAlignedSuperStructure(pdb.location, assembly.location)

#Perform Analysis
#We use a very high alpha level here with no multiple comparison adjustment
#to make sure that each method provides shows a result.
#Lower alpha cut offs are typically used.
(quart_results <- quartCluster(mutation.data, structural.data, perform.ipac = "Y", perform.graphpac = "Y",
                              perform.spacepac = "Y", create.map = "N", MultComp = "None",
                              alpha = .3, radii.vector = c(1:3), show.low.level.messages = "Y"))
}
}
\keyword{protein}
\keyword{quarternary}
\keyword{clustering}
